requirements.txt
sentence-transformers
faiss-cpu
requests
numpy
pydantic-settings
python-dotenv